refactor(estimation): sort calculation code
